% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.data.frame.tmbfit}
\alias{as.data.frame.tmbfit}
\title{Convert object of class tmbfit to data.frame. Calls
\code{\link{extract_samples}}}
\usage{
\method{as.data.frame}{tmbfit}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Fitted object from \code{\link{sample_snuts}}}

\item{row.names}{Ignored}

\item{optional}{Ignored}

\item{...}{Ignored}
}
\value{
A data frame with parameters as columns and samples as
rows.
}
\description{
Convert object of class tmbfit to data.frame. Calls
\code{\link{extract_samples}}
}
\details{
This calls the default settings of
\code{\link{extract_samples}}, no warmup samples and no
column for the log-posterior (lp__). Use this function
directly for finer control.
}
