% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_snuts_diagnostics}
\alias{check_snuts_diagnostics}
\title{Check NUTS diagnostics of a fitted model}
\usage{
check_snuts_diagnostics(fit, print = TRUE)
}
\arguments{
\item{fit}{A fitted SNUTS object}

\item{print}{Whether to print the results to console}
}
\value{
A data.frame containing diagnostic results (invisibly)
}
\description{
Check NUTS diagnostics of a fitted model
}
\details{
This is a wrapper which calls the internal function \code{\link[StanEstimators]{check_hmc_diagnostics}}
}
