% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initalization.R
\name{.get_inits}
\alias{.get_inits}
\title{Get a single initial value vector in untransformed model space}
\usage{
.get_inits(init, obj2, inputs)
}
\arguments{
\item{init}{The initial value strategy}

\item{obj2}{The joint TMB model}

\item{inputs}{A list as returned by \code{.get_inputs}.}
}
\description{
Get a single initial value vector in untransformed model space
}
