% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precondition.R
\name{.precondition}
\alias{.precondition}
\title{Update algorithm for mass matrix.}
\usage{
.precondition(metric, fn, gr, inputs, y.cur)
}
\arguments{
\item{metric}{The metric to use}

\item{fn}{The current fn function.}

\item{gr}{The current gr function}

\item{inputs}{A list of inputs}

\item{y.cur}{The current parameter vector in unrotated (Y) space.}
}
\description{
Update algorithm for mass matrix.
}
