% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precondition.R
\name{.rotate_posterior}
\alias{.rotate_posterior}
\title{Update algorithm for mass matrix.}
\usage{
.rotate_posterior(metric, fn, gr, inputs, y.cur)
}
\arguments{
\item{metric}{The metric to use}

\item{fn}{The current fn function.}

\item{gr}{The current gr function}

\item{y.cur}{The current parameter vector in unrotated (Y) space.}

\item{Q}{The sparse precision matrix}

\item{Qinv}{The inverse of Q}
}
\description{
Update algorithm for mass matrix.
}
