% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sparse_correlation}
\alias{get_sparse_correlation}
\title{Extract a single correlation from a sparse precision matrix}
\usage{
get_sparse_correlation(Q, i, j)
}
\arguments{
\item{Q}{The sparse precision}

\item{i, j}{The integer indices for the correlation}
}
\value{
A single scalar correlation value
}
\description{
Extract a single correlation from a sparse precision matrix
}
\details{
This function extracts correlations without having to
invert Q using sparse matrix routines. Used for plotting outputs only.
}
