% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.snutsfit}
\alias{is.snutsfit}
\title{Check object of class snutsfit}
\usage{
is.snutsfit(x)
}
\arguments{
\item{x}{Returned list from \code{\link{sample_admb}}}
}
\description{
Check object of class snutsfit
}
