% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.R
\name{pairs.tmbfit}
\alias{pairs.tmbfit}
\title{Plot pairwise parameter posteriors and optionally the MLE points and
confidence ellipses.}
\usage{
\method{pairs}{tmbfit}(
  x,
  pars = NULL,
  order = c("orig", "slow", "fast", "mismatch", "cor"),
  inc_warmup = FALSE,
  diag = c("trace", "acf", "hist"),
  acf.ylim = c(-1, 1),
  ymult = NULL,
  axis.col = gray(0.5),
  label.cex = 0.8,
  limits = NULL,
  add.mle = TRUE,
  add.monitor = TRUE,
  add.inits = FALSE,
  point.col = NULL,
  point.pch = NULL,
  plot = TRUE,
  unbounded = FALSE,
  ...
)
}
\arguments{
\item{x}{A list as returned by \code{sample_nuts}.}

\item{pars}{A character vector of parameters or integers
representing which parameters to subset. Useful if the model
has a larger number of parameters and you just want to show
a few key ones.}

\item{order}{The order to consider the parameters. Options are
'orig' (default) to use the order declared in the model, or
'slow' and 'fast' which are based on the effective sample
sizes ordered by slowest or fastest mixing
respectively. 'mismatch' sorts by parameters with
large discrepancies between the MLE and posterior marginal
variances, defined as the absolute relative difference of
the MLE from the posterior i.e., abs((mle-post)/post).
Finally, 'cor' orders by the largest maximum absolute
pairwise posterior correlation (including lp__).
See example for usage.}

\item{inc_warmup}{Whether to include the warmup samples or not
(default).}

\item{diag}{What type of plot to include on the diagonal,
options are 'acf' which plots the autocorrelation function
\code{acf}, 'hist' shows marginal posterior histograms, and
'trace' the trace plot.}

\item{acf.ylim}{If using the acf function on the diagonal,
specify the y limit. The default is c(-1,1).}

\item{ymult}{A vector of length ncol(posterior) specifying how
much room to give when using the hist option for the
diagonal. For use if the label is blocking part of the
plot. The default is 1.3 for all parameters.}

\item{axis.col}{Color of axes}

\item{label.cex}{Control size of outer and diagonal labels
(default 1)}

\item{limits}{A list containing the ranges for each parameter
to use in plotting.}

\item{add.mle}{Boolean whether to add 95\% confidence ellipses}

\item{add.monitor}{Boolean whether to print effective sample}

\item{add.inits}{Boolean whether to add the initial values to the plot}

\item{point.col}{Color of posterior points. Default selects level of transparency depending on the number of samples.}

\item{point.pch}{Shape of posterior points. Defaults to 16 if unspecified.}

\item{plot}{Whether to plot (default) the results. If FALSE it will return the parameter names.}

\item{unbounded}{Whether to use the bounded or unbounded
version of the parameters.  size (ESS) and Rhat values on
the diagonal.}

\item{...}{Arguments to be passed to plot call in lower
triangular panels (scatterplots).}
}
\value{
Produces a plot, and returns nothing by default. If
argument \code{plot} is FALSE then it will return a named
character vector which would have been plotted.
}
\description{
Plot pairwise parameter posteriors and optionally the MLE points and
confidence ellipses.
}
\details{
This function is modified from the base \code{pairs}
code to work specifically with fits from the \code{\link{sample_snuts}}
function using the SNUTS algorithms. If an
invertible Hessian was found (in \code{fit$mle}) then
estimated covariances are available to compare and added
automatically (red ellipses). Likewise, a "monitor" object
from \code{\link[posterior]{summarize_draws}} is attached as \code{fit$monitor}
and provides effective sample sizes (ESS) and Rhat
values. The ESS are used to potentially order the parameters
via argument \code{order}, but also printed on the diagonal.
}
\examples{
fit <- readRDS(system.file('examples', 'fit.RDS', package='SparseNUTS'))
pairs(fit)
pairs(fit, pars=1:2)
pairs(fit, pars=c(2,1))
pairs(fit, pars=1:2, order='slow')
pairs(fit, pars=1:2, order='fast')
pairs(fit, pars=1:2, order='mismatch')

}
\author{
Cole Monnahan
}
