% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.snutsfit}
\alias{plot.snutsfit}
\title{Plot object of class snutsfit}
\usage{
\method{plot}{snutsfit}(x, y, ...)
}
\arguments{
\item{x}{Fitted object from \code{\link{sample_admb}}}

\item{y}{Ignored}

\item{...}{Ignored}
}
\value{
Plot created
}
\description{
Plot object of class snutsfit
}
