% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.tmbfit}
\alias{plot.tmbfit}
\title{Plot object of class tmbfit}
\usage{
\method{plot}{tmbfit}(x, y, ...)
}
\arguments{
\item{x}{Fitted object from \code{\link{sample_snuts}}}

\item{y}{Ignored}

\item{...}{Ignored}
}
\value{
Plot created
}
\description{
Plot object of class tmbfit
}
