% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.snutsfit}
\alias{print.snutsfit}
\title{Print summary of snutsfit object}
\usage{
\method{print}{snutsfit}(x, ...)
}
\arguments{
\item{x}{Fitted object from \code{\link{sample_admb}}}

\item{...}{Ignored}
}
\value{
Summary printed to console
}
\description{
Print summary of snutsfit object
}
