% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.tmbfit}
\alias{print.tmbfit}
\title{Print summary of tmbfit object}
\usage{
\method{print}{tmbfit}(x, ...)
}
\arguments{
\item{x}{Fitted object from \code{\link{sample_snuts}}}

\item{...}{Ignored}
}
\value{
Summary printed to console
}
\description{
Print summary of tmbfit object
}
