% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initalization.R
\name{rmvnorm_Q}
\alias{rmvnorm_Q}
\title{Simulate draws from a MVN using the precision matrix Q}
\usage{
rmvnorm_Q(Q, nsim)
}
\arguments{
\item{Q}{A sparse precision matrix}

\item{nsim}{How many replicates to simulate}
}
\description{
Simulate draws from a MVN using the precision matrix Q
}
