% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initalization.R
\name{sample_inits}
\alias{sample_inits}
\title{Function to generate random initial values from a previous fit}
\usage{
sample_inits(fit, chains)
}
\arguments{
\item{fit}{A fitted object from \code{\link{sample_snuts}}}

\item{chains}{The number of chains for the subsequent run, which
determines the number to return.}
}
\value{
A list of vectors which can be passed back into
\code{\link{sample_snuts}}.
}
\description{
Function to generate random initial values from a previous fit
}
