% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snutsfit}
\alias{snutsfit}
\title{Constructor for the "snutsfit" (A-D fit) class}
\usage{
snutsfit(x)
}
\arguments{
\item{x}{Fitted object from \code{\link{sample_admb}}}
}
\value{
An object of class "snutsfit"
}
\description{
Constructor for the "snutsfit" (A-D fit) class
}
