% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.snutsfit}
\alias{summary.snutsfit}
\title{Print summary of object of class snutsfit}
\usage{
\method{summary}{snutsfit}(object, ...)
}
\arguments{
\item{object}{Fitted object from \code{\link{sample_admb}}}

\item{...}{Ignored}
}
\value{
Summary printed to screen
}
\description{
Print summary of object of class snutsfit
}
