% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.tmbfit}
\alias{summary.tmbfit}
\title{Print summary of object of class tmbfit}
\usage{
\method{summary}{tmbfit}(object, ...)
}
\arguments{
\item{object}{Fitted object from \code{\link{sample_snuts}}}

\item{...}{Ignored}
}
\value{
Summary printed to screen
}
\description{
Print summary of object of class tmbfit
}
