% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tmbfit}
\alias{tmbfit}
\title{Constructor for the "tmbfit"  class}
\usage{
tmbfit(x)
}
\arguments{
\item{x}{Fitted object from \code{\link{sample_snuts}}}
}
\value{
An object of class "tmbfit"
}
\description{
Constructor for the "tmbfit"  class
}
