% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.print.mat.stats}
\alias{.print.mat.stats}
\title{Print matrix stats}
\usage{
.print.mat.stats(stats, getCF)
}
\arguments{
\item{stats}{A list of matrix stats}

\item{getCF}{A flag whether to calculate the HMC condition factor.}
}
\value{
Prints the HMC condition factor, sparsity, and max absolute correlation
}
\description{
Print matrix stats
}
\details{
The HMC condition factor used is from Langmore et al.
(2019). A NULL value is returned if the matrix passed is a
scalar or if getCF=FALSE.
}
