% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initalization.R
\name{sample_Q}
\alias{sample_Q}
\title{Generate approximate posterior samples from the sparse
precision matrix Q, assuming multivariate normality.}
\usage{
sample_Q(fit, nsamples = 4000, seed = NULL)
}
\arguments{
\item{fit}{An object returned by \code{sample_snuts}}

\item{nsamples}{The number of samples to draw, defaulting to 4000.}

\item{seed}{An optional random seed}
}
\value{
A data.frame of approximate samples
}
\description{
Generate approximate posterior samples from the sparse
precision matrix Q, assuming multivariate normality.
}
\examples{
fit <- readRDS(system.file('examples', 'fit.RDS', package='SparseNUTS'))
x <- sample_Q(fit)
}
